service-authentication/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
setuptools